concluded parsing the update action incorrectly
Initially I thought it was re-using the objects, but sending ALL the
data.

Now I think it's re-using the objects and only sending a DIFF of the
data.

+++ protocol-java-impl/src/rotmg/util/BitsAndBytes.java
@@ -39,7 +39,7 @@ public class BitsAndBytes {
     }
     
     public static List<Integer> readIntList(DataInputStream din) throws IOException {
-        int size = din.readInt();
+        int size = din.readShort();
         
         List<Integer> list = new ArrayList<>();
         for(int i = 0; i < size; i++) {
